refactor: JWT 관련 class refactoring
