+++ perfume-core/src/main/java/jwt/JsonWebTokenGenerator.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Component;
 import java.security.Key;
 import java.time.LocalDateTime;
 import java.util.Date;
-import java.util.List;
 import java.util.Map;
 
 @Component
