+++ perfume-core/src/test/java/jwt/JsonWebTokenGeneratorTest.java
@@ -2,7 +2,6 @@ package jwt;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.platform.commons.util.StringUtils;
 
 import java.time.LocalDateTime;
 import java.util.HashMap;
